Rename hand and move painter
