+++ app/lib/views/view.dart
@@ -217,7 +217,7 @@ class _MainViewViewportState extends State<MainViewViewport>
                           }
                           if (openView) openView = details.scale == 1;
                           final settings = context.read<SettingsCubit>().state;
-                          if (cubit.fetchHandler<HandHandler>() == null &&
+                          if (cubit.fetchHandler<SelectHandler>() == null &&
                               !settings.inputGestures) return;
                           var current = details.scale;
                           current = current - size;
