+++ docs/package.json
@@ -24,7 +24,7 @@
     "@docusaurus/preset-classic": "2.4.1",
     "@docusaurus/theme-common": "2.4.1",
     "@mdx-js/react": "^1",
-    "@swc/core": "^1.3.76",
+    "@swc/core": "^1.3.77",
     "animate.css": "^4.1.1",
     "clsx": "^2.0.0",
     "node-fetch": "^3.3.2",
